Move configuration logics to the Configuration module
